Welcome to v2!
After some benchmarks made it clear there was vast room for improvement,
the code was torn down and built back up from scratch. In the end, it
seemed best to reuse existing code from bytes.Buffer and bytes.Reader
to provide the underpinnings, and then built up from there. Given
the large amount of code that was BSD-licensed from Go, it made sense
to relicense the MIT-side we held to BSD as well, so dual-licensing
and pedantic provenance wasn't critical.